# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Simula Native Ad Server",
        "description": "Campaign/ad-set management, session resolution and native ad serving with geo + OS targeting, bandit ranking and LLM ad copy.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/campaigns": {
            "post": {
                "tags": [
                    "campaigns"
                ],
                "summary": "Create Campaign",
                "operationId": "create_campaign_campaigns_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campaign"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "campaigns"
                ],
                "summary": "List Campaigns",
                "operationId": "list_campaigns_campaigns_get",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "comma-separated campaign ids",
                            "title": "Ids"
                        },
                        "description": "comma-separated campaign ids"
                    },
                    {
                        "name": "surface",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Surface"
                        }
                    },
                    {
                        "name": "publisher_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Publisher Id"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Active"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Campaign"
                                    },
                                    "title": "Response List Campaigns Campaigns Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/campaigns/{campaign_id}": {
            "get": {
                "tags": [
                    "campaigns"
                ],
                "summary": "Get Campaign",
                "operationId": "get_campaign_campaigns__campaign_id__get",
                "parameters": [
                    {
                        "name": "campaign_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Campaign Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campaign"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "campaigns"
                ],
                "summary": "Update Campaign",
                "operationId": "update_campaign_campaigns__campaign_id__patch",
                "parameters": [
                    {
                        "name": "campaign_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Campaign Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campaign"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "campaigns"
                ],
                "summary": "Delete Campaign",
                "operationId": "delete_campaign_campaigns__campaign_id__delete",
                "parameters": [
                    {
                        "name": "campaign_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Campaign Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/load/native": {
            "post": {
                "tags": [
                    "serving"
                ],
                "summary": "Load Native",
                "description": "Serve a native sponsored-character ad for a feed slot.",
                "operationId": "load_native_load_native_post",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "pattern": "^(dark|light)$",
                            "default": "dark",
                            "title": "Theme"
                        }
                    },
                    {
                        "name": "X-Forwarded-For",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Forwarded-For"
                        }
                    },
                    {
                        "name": "X-Real-IP",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Real-Ip"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "User-Agent"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NativeAdRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NativeAdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/impressions/{impression_id}/click": {
            "post": {
                "tags": [
                    "serving"
                ],
                "summary": "Record Click",
                "description": "Click callback fired by the rendered creative.\n\nThe template calls this endpoint directly, so it exists even though the\nbrief does not list it. Returns 202 on an unknown or already-counted id\nrather than 404: click pixels are fire-and-forget and retried by the\nbrowser, and a duplicate must never look like a failure the client should\nretry harder.",
                "operationId": "record_click_impressions__impression_id__click_post",
                "parameters": [
                    {
                        "name": "impression_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Impression Id"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Record Click Impressions  Impression Id  Click Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Campaign": {
                "properties": {
                    "campaign_id": {
                        "type": "string",
                        "title": "Campaign Id"
                    },
                    "campaign_name": {
                        "type": "string",
                        "title": "Campaign Name"
                    },
                    "advertiser_company_id": {
                        "type": "string",
                        "title": "Advertiser Company Id"
                    },
                    "active": {
                        "type": "boolean",
                        "title": "Active",
                        "default": false
                    },
                    "daily_budget": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Daily Budget"
                    },
                    "geo_targets": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Geo Targets"
                    },
                    "os_targets": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Os Targets"
                    },
                    "attribution_provider": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attribution Provider"
                    },
                    "ios_store_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ios Store Url"
                    },
                    "android_store_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Android Store Url"
                    },
                    "native_ad_set_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Native Ad Set Ids"
                    },
                    "surface": {
                        "type": "string",
                        "title": "Surface",
                        "default": "native"
                    },
                    "publisher_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Publisher Id"
                    },
                    "impression_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Impression Url"
                    },
                    "downloads_label": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Downloads Label"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "campaign_name",
                    "advertiser_company_id"
                ],
                "title": "Campaign"
            },
            "CampaignCreate": {
                "properties": {
                    "campaign_name": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 1,
                        "title": "Campaign Name"
                    },
                    "advertiser_company_id": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Advertiser Company Id"
                    },
                    "daily_budget": {
                        "anyOf": [
                            {
                                "type": "number",
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Daily Budget"
                    },
                    "geo_targets": {
                        "items": {
                            "type": "string",
                            "maxLength": 2,
                            "minLength": 2,
                            "pattern": "^[A-Za-z]{2}$"
                        },
                        "type": "array",
                        "title": "Geo Targets"
                    },
                    "os_targets": {
                        "items": {
                            "$ref": "#/components/schemas/OS"
                        },
                        "type": "array",
                        "title": "Os Targets"
                    },
                    "attribution_provider": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attribution Provider"
                    },
                    "ios_store_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2083,
                                "minLength": 1,
                                "format": "uri"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ios Store Url"
                    },
                    "android_store_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2083,
                                "minLength": 1,
                                "format": "uri"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Android Store Url"
                    },
                    "native_ad_set_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Native Ad Set Ids"
                    },
                    "surface": {
                        "$ref": "#/components/schemas/Surface",
                        "default": "native"
                    },
                    "publisher_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Publisher Id"
                    },
                    "impression_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2083,
                                "minLength": 1,
                                "format": "uri"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Impression Url"
                    },
                    "downloads_label": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Downloads Label"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "campaign_name",
                    "advertiser_company_id"
                ],
                "title": "CampaignCreate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "NativeAdRequest": {
                "properties": {
                    "position": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Position",
                        "description": "feed index of the slot"
                    },
                    "session_id": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Session Id"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AdContext"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "position",
                    "session_id"
                ],
                "title": "NativeAdRequest"
            },
            "NativeAdResponse": {
                "properties": {
                    "impression_id": {
                        "type": "string",
                        "title": "Impression Id"
                    },
                    "rendered_html": {
                        "type": "string",
                        "title": "Rendered Html"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "impression_id",
                    "rendered_html"
                ],
                "title": "NativeAdResponse"
            },
...
        }
    }
}